pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-debug-env"
version = "1.0.0"
description = "SQL Debug & Data Pipeline Repair — OpenEnv environment with Four tasks"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
keywords = ["openenv", "reinforcement-learning", "sql", "duckdb", "data-engineering"]

dependencies = [
    "duckdb>=0.10.0",
    "pandas>=2.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "openai>=1.30.0",
    "pyyaml>=6.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
openenv = [
    "openenv-core>=0.1.0",
]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sql_env*", "server*"]